+++ tests/namespaces/ta/oscillators-momentum.test.ts
@@ -607,4 +607,7 @@ describe('Technical Analysis - Oscillators & Momentum', () => {
         console.log('plotdata_str', plotdata_str);
         expect(plotdata_str.trim()).toEqual(expected_plot.trim());
     });
+
+
+       
 });